app/requirements.txt
# Core web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Authentication and security
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
pyotp>=2.9.0
cryptography>=44.0.0

# Data processing
pydantic>=2.11.0
python-multipart>=0.0.20
python-dotenv>=1.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Trading related
python-binance>=1.0.28
requests>=2.32.0

# Data analysis
pandas>=2.2.0
numpy>=1.24.0

# Redis
redis>=6.0.0

# Other utilities
python-dateutil>=2.9.0
pytz>=2025.0

# Additional dependencies for project
qrcode[pil]
psutil
